added decorators/ tests file added
So I added a decorators, changed order of the code, and added extra files with a tests.

I have a troubles with adding more decorators for updating and removing a file. And I mean code is not working properly even when I am adding a new property with different values.